requirements.txt
attrs==19.3.0
Click==7.0
coverage==4.5.4
Flask==1.1.1
gunicorn==19.9.0
importlib-metadata==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.0
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
six==1.13.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
 
flask-restful
flask-restful-swagger-2
flask-swagger-ui
Flask-Cors
python-dotenv==0.12.0
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
psycopg2-binary


